+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The API is unstable, and [Semantic Versioning](https://semver.org/spec/v2.0.0.ht
 ### Added
 - `MenuBar` method `closeMenus()` to close any open menus.
 - `MenuBar` method `setKeyboardScope(...elements)` to control hotkey handling
+- `$Window` method `setMenuBar(menuBar)` to set the menu bar, and set up the keyboard scope.
 - `$Window` property `element` to get the DOM element.
 - `element.$window` to get the `$Window` instance from the DOM element.